+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Tracing: span links on receive/send spans now fall back to using `Diagnostic-Id` if the `traceparent` message application property is not found.
+
 ## 7.10.0 (2023-05-09)
 
 Version 7.10.0 is our first stable release of the Azure Service Bus client library based on a pure Python implemented AMQP stack.